# `RegexFSM`: Cache Legal-Token Mask as `torch.Tensor` to Improve Performance