--- a/tests/fsm/test_fsm.py
+++ b/tests/fsm/test_fsm.py
@@ -3,6 +3,13 @@ import pytest
 from outlines.fsm.fsm import CFGFSM, RegexFSM, StopAtEosFSM
 
 
+def assert_expected_tensor_ids(tensor, ids):
+    assert len(tensor) == len(ids)
+    norm_tensor = sorted(map(int, tensor))
+    norm_ids = sorted(map(int, tensor))
+    assert norm_tensor == norm_ids, (norm_tensor, norm_ids)
+
+
 def test_stop_at_eos():
     class MockTokenizer:
         vocabulary = {"a": 1, "eos": 2}
@@ -50,7 +57,7 @@ def test_regex():
         fsm = RegexFSM(regex_str, tokenizer)
 
     assert fsm.states_to_token_maps == {0: {1: 1}}
-    assert fsm.allowed_token_ids(state=0) == [1]
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=0), [1])
     assert fsm.next_state(state=0, token_id=1) == 1
     assert fsm.next_state(state=0, token_id=tokenizer.eos_token_id) == -1
 
@@ -111,27 +118,27 @@ def test_cfg():
     with pytest.warns(UserWarning):
         fsm = CFGFSM(cfg_str, tokenizer)
 
-    assert set(fsm.allowed_token_ids(state=fsm.start_state)) == {1, 3, 5}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=fsm.start_state), [1, 3, 5])
     state = fsm.next_state(state=fsm.start_state, token_id=1)
     assert fsm.generation == "{"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 2, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 2, 3])
     state = fsm.next_state(state=state, token_id=3)
     assert fsm.generation == "{["
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 3, 4}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 3, 4])
     state = fsm.next_state(state=state, token_id=4)
     assert fsm.generation == "{[]"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {2}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [2])
     state = fsm.next_state(state=state, token_id=2)
     assert fsm.generation == "{[]}"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {5}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [5])
     state = fsm.next_state(state=state, token_id=5)
     assert fsm.generation == "{[]}"
     assert fsm.is_final_state(state)
@@ -164,24 +171,24 @@ def test_cfg_early_termination():
     with pytest.warns(UserWarning):
         fsm = CFGFSM(cfg_str, tokenizer)
 
-    assert set(fsm.allowed_token_ids(state=fsm.start_state)) == {1}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=fsm.start_state), [1])
     state = fsm.next_state(state=fsm.start_state, token_id=1)
     assert fsm.generation == "("
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 2}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 2])
     state = fsm.next_state(state=state, token_id=2)
     assert fsm.generation == "()"
     assert not fsm.is_final_state(state)
 
     # possible to continue or terminate
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 3])
     state = fsm.next_state(state=state, token_id=3)  # feed eos
     assert fsm.generation == "()"
     assert fsm.is_final_state(state)
 
     # once eos generated, can only terminate
-    assert set(fsm.allowed_token_ids(state=state)) == {3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [3])
 
 
 def test_cfg_ignore_directive():
@@ -214,38 +221,38 @@ def test_cfg_ignore_directive():
 
     state = 0
 
-    assert set(fsm.allowed_token_ids(state=0)) == {1, 2}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=0), [1, 2])
     state = fsm.next_state(state=0, token_id=2)
     assert fsm.generation == " "
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=0)) == {1, 2}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=0), [1, 2])
     state = fsm.next_state(state=0, token_id=1)
     assert fsm.generation == " a"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 2, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 2, 3])
     state = fsm.next_state(state=state, token_id=2)
     assert fsm.generation == " a "
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 2, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 2, 3])
     state = fsm.next_state(state=state, token_id=2)
     assert fsm.generation == " a  "
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 2, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 2, 3])
     state = fsm.next_state(state=state, token_id=1)
     assert fsm.generation == " a  a"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 2, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 2, 3])
     state = fsm.next_state(state=state, token_id=3)
     assert fsm.generation == " a  a"
     assert fsm.is_final_state(state)
 
     # once eos generated, can only terminate
-    assert set(fsm.allowed_token_ids(state=state)) == {3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [3])
 
 
 def test_cfg_multitoken_terminal():
@@ -274,19 +281,19 @@ def test_cfg_multitoken_terminal():
     with pytest.warns(UserWarning):
         fsm = CFGFSM(cfg_str, tokenizer)
 
-    assert set(fsm.allowed_token_ids(state=fsm.start_state)) == {1, 2}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=fsm.start_state), [1, 2])
     assert fsm.reset_state  # starting new regex
     state = fsm.next_state(state=fsm.start_state, token_id=1)
     assert fsm.generation == "a"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1])
     assert not fsm.reset_state  # continuing current regex
     state = fsm.next_state(state=state, token_id=1)
     assert fsm.generation == "aa"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [3])
     assert not fsm.reset_state  # completing current regex
     state = fsm.next_state(state=state, token_id=3)
     assert fsm.generation == "aa"
@@ -319,27 +326,27 @@ def test_cfg_allow_both_extend_and_shift_terminal():
     with pytest.warns(UserWarning):
         fsm = CFGFSM(cfg_str, tokenizer)
 
-    assert set(fsm.allowed_token_ids(state=fsm.start_state)) == {1, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=fsm.start_state), [1, 3])
     state = fsm.next_state(state=fsm.start_state, token_id=1)
     assert fsm.generation == "("
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {1, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [1, 3])
     state = fsm.next_state(state=state, token_id=3)
     assert fsm.generation == "(a"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {2, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [2, 3])
     state = fsm.next_state(state=state, token_id=3)
     assert fsm.generation == "(aa"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {2, 3}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [2, 3])
     state = fsm.next_state(state=state, token_id=2)
     assert fsm.generation == "(aa)"
     assert not fsm.is_final_state(state)
 
-    assert set(fsm.allowed_token_ids(state=state)) == {4}
+    assert_expected_tensor_ids(fsm.allowed_token_ids(state=state), [4])
     state = fsm.next_state(state=state, token_id=4)
     assert fsm.generation == "(aa)"
     assert fsm.is_final_state(state)