@lapp0 2024-07-02
@Qubitium yes, the gains are proportional to the size of the tokenizers vocabulary.